pyproject.toml
# 打包配置
[tool.setuptools.packages.find]
where = ["."]  # 指定包目录
include = ["utils", "testCases"]  # 明确包含的包

# 测试配置
[tool.pytest.ini_options]
testpaths = ["testCases"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=61.0", "wheel"]

[project]
name = "auto-tester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "allure-pytest>=2.14.3",
    "jsonpath>=0.82.2",
    "pymysql>=1.1.1",
    "pytest>=8.4.1",
    "pytest-html>=4.1.1",
    "pytest-result-log>=1.2.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
